moved `Item` to item.py module
